fix: all responses are json
